custom_yolo/requirements.txt
torch
torchvision
pillow
scikit-learn
matplotlib
seaborn
numpy
wandb
kaggle
